Make getBets a hook to return hooks
