+++ frontend/package.json
@@ -31,6 +31,10 @@
   "eslintConfig": {
     "extends": "react-app"
   },
+  "resolutions": {
+    "**/@typescript-eslint/eslint-plugin": "^4.1.1",
+    "**/@typescript-eslint/parser": "^4.1.1"
+  },
   "browserslist": {
     "production": [
       ">0.2%",
